Rewrite library structure to more professional
Using github.com/futuresimple/dropbox-api as a model, Clfu::API::Client
now handles pulling data from the service with instances instead of
using class methods as before. One obvious advantage of doing it this
way is that instances will be able to cache separately. Last, the API
module exists because a CLI module is planned in the future.